+++ Runtime/iOS/Plugins/iOS/Utilities/CBCUnityUtilities.h
@@ -6,6 +6,7 @@
 typedef void (*ChartboostCoreOnModuleInitializationResult)(const char* moduleIdentifier, long start, long end, long duration, const char* exception);
 typedef void (*ChartboostCoreOnModuleInitializeDelegate)(const char* moduleIdentifier, const char* chartboostAppIdentifier);
 typedef void (*ChartboostCoreOnConsentChangeForStandard)(const char* standard, const char* value);
+typedef void (*ChartboostCoreOnPartnerConsentChange)(const char* partnerIdentifier, int consentStatus);
 typedef void (*ChartboostCoreOnEnumStatusChange)(int value);
 typedef void (*ChartbosotCoreResultString)(int hashCode, const char* result);
 typedef void (*ChartboostCoreResultBoolean)(int hashCode, bool completion);